Add files via upload
Context and Obejective:

The dataset being used here has 7 years' historical loan data taken from LendingClub. By looking at the loan status, we can check if the loan was sanctioned to the customers or not. If the loan status is "charged off" or "default" or the issuance of loan got delayed by 31-120 days, it is a bad loan. We will analyze the characteristics of the customers whose loans belong to "bad loan" category and will build a Probability of Default (PD) Model, so that is can be used in future to predict the customer will default or not.

We have used some useful concepts:

- Binning,
- Weight of Evidence (WoE),
- Information Value (IV)